[Internal] Refactor provider packages
