+++ internal/common/provider/provider_test.go
@@ -12,6 +12,7 @@ import (
 	"time"
 
 	"github.com/databricks/terraform-provider-databricks/common"
+	sdkv2provider "github.com/databricks/terraform-provider-databricks/internal/sdkv2/provider"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -179,7 +180,7 @@ func TestConfig_PatFromDatabricksCfg(t *testing.T) {
 	providerFixture{
 		// loading with DEFAULT profile in databrickscfs
 		env: map[string]string{
-			"HOME": "../common/testdata",
+			"HOME": "../../../common/testdata",
 		},
 		assertHost: "https://dbc-XXXXXXXX-YYYY.cloud.databricks.com",
 		assertAuth: "pat",
@@ -190,7 +191,7 @@ func TestConfig_PatFromDatabricksCfg_NohostProfile(t *testing.T) {
 	providerFixture{
 		// loading with nohost profile in databrickscfs
 		env: map[string]string{
-			"HOME":                      "../common/testdata",
+			"HOME":                      "../../../common/testdata",
 			"DATABRICKS_CONFIG_PROFILE": "nohost",
 		},
 		assertError: common.NoAuth +
@@ -203,7 +204,7 @@ func TestConfig_ConfigProfileAndToken(t *testing.T) {
 		env: map[string]string{
 			"DATABRICKS_TOKEN":          "x",
 			"DATABRICKS_CONFIG_PROFILE": "nohost",
-			"HOME":                      "../common/testdata",
+			"HOME":                      "../../../common/testdata",
 		},
 		assertError: common.NoAuth +
 			". Config: token=***, profile=nohost. Env: DATABRICKS_TOKEN, DATABRICKS_CONFIG_PROFILE",
@@ -215,7 +216,7 @@ func TestConfig_ConfigProfileAndPassword(t *testing.T) {
 		env: map[string]string{
 			"DATABRICKS_USERNAME":       "x",
 			"DATABRICKS_CONFIG_PROFILE": "nohost",
-			"HOME":                      "../common/testdata",
+			"HOME":                      "../../../common/testdata",
 		},
 		assertError: "validate: more than one authorization method configured: basic and pat. " +
 			"Config: token=***, username=x, profile=nohost. Env: DATABRICKS_USERNAME, DATABRICKS_CONFIG_PROFILE",
@@ -225,7 +226,7 @@ func TestConfig_ConfigProfileAndPassword(t *testing.T) {
 var azResourceID = "/subscriptions/a/resourceGroups/b/providers/Microsoft.Databricks/workspaces/c"
 
 func TestConfig_AzureCliHost(t *testing.T) {
-	p, _ := filepath.Abs("../common/testdata")
+	p, _ := filepath.Abs("../../../common/testdata")
 	providerFixture{
 		// this test will skip ensureWorkspaceUrl
 		host:            "x",
@@ -243,7 +244,7 @@ func TestConfig_AzureCliHost(t *testing.T) {
 }
 
 func TestConfig_AzureCliHost_Fail(t *testing.T) {
-	p, _ := filepath.Abs("../common/testdata")
+	p, _ := filepath.Abs("../../../common/testdata")
 	providerFixture{
 		azureResourceID: azResourceID,
 		env: map[string]string{
@@ -262,14 +263,14 @@ func TestConfig_AzureCliHost_AzNotInstalled(t *testing.T) {
 		azureResourceID: azResourceID,
 		env: map[string]string{
 			"PATH": "whatever",
-			"HOME": "../common/testdata",
+			"HOME": "../../common/testdata",
 		},
 		assertError: common.NoAuth,
 	}.apply(t)
 }
 
 func TestConfig_AzureCliHost_PatConflict(t *testing.T) {
-	p, _ := filepath.Abs("../common/testdata")
+	p, _ := filepath.Abs("../../../common/testdata")
 	providerFixture{
 		azureResourceID: azResourceID,
 		token:           "x",
@@ -283,7 +284,7 @@ func TestConfig_AzureCliHost_PatConflict(t *testing.T) {
 }
 
 func TestConfig_AzureCliHostAndResourceID(t *testing.T) {
-	p, _ := filepath.Abs("../common/testdata")
+	p, _ := filepath.Abs("../../../common/testdata")
 	providerFixture{
 		// omit request to management endpoint to get workspace properties
 		azureResourceID: azResourceID,
@@ -301,7 +302,7 @@ func TestConfig_AzureCliHostAndResourceID(t *testing.T) {
 }
 
 func TestConfig_AzureAndPasswordConflict(t *testing.T) {
-	p, _ := filepath.Abs("../common/testdata")
+	p, _ := filepath.Abs("../../../common/testdata")
 	providerFixture{
 		host:            "x",
 		azureResourceID: azResourceID,
@@ -318,7 +319,7 @@ func TestConfig_AzureAndPasswordConflict(t *testing.T) {
 func TestConfig_CorruptConfig(t *testing.T) {
 	providerFixture{
 		env: map[string]string{
-			"HOME": "../common/testdata/corrupt",
+			"HOME": "../../../common/testdata/corrupt",
 		},
 		assertError: common.NoAuth,
 	}.apply(t)
@@ -390,7 +391,7 @@ type parseUserAgentTestCase struct {
 	name string
 	env  string
 	err  error
-	out  []userAgentExtra
+	out  []sdkv2provider.UserAgentExtra
 }
 
 func Test_ParseUserAgentExtra(t *testing.T) {
@@ -399,7 +400,7 @@ func Test_ParseUserAgentExtra(t *testing.T) {
 			name: "single product",
 			env:  "databricks-cli/0.1.2",
 			err:  nil,
-			out: []userAgentExtra{
+			out: []sdkv2provider.UserAgentExtra{
 				{"databricks-cli", "0.1.2"},
 			},
 		},
@@ -407,7 +408,7 @@ func Test_ParseUserAgentExtra(t *testing.T) {
 			name: "multiple products",
 			env:  "databricks-cli/0.1.2 custom-thing/0.0.1",
 			err:  nil,
-			out: []userAgentExtra{
+			out: []sdkv2provider.UserAgentExtra{
 				{"databricks-cli", "0.1.2"},
 				{"custom-thing", "0.0.1"},
 			},
@@ -416,7 +417,7 @@ func Test_ParseUserAgentExtra(t *testing.T) {
 			name: "multiple products with many separators",
 			env:  "\ta/0.0.1\tb/0.0.2 \t c/0.0.3",
 			err:  nil,
-			out: []userAgentExtra{
+			out: []sdkv2provider.UserAgentExtra{
 				{"a", "0.0.1"},
 				{"b", "0.0.2"},
 				{"c", "0.0.3"},
@@ -426,7 +427,7 @@ func Test_ParseUserAgentExtra(t *testing.T) {
 			name: "empty string",
 			env:  "",
 			err:  nil,
-			out:  []userAgentExtra{},
+			out:  []sdkv2provider.UserAgentExtra{},
 		},
 		{
 			name: "product with comment",
@@ -450,7 +451,7 @@ func Test_ParseUserAgentExtra(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			out, err := parseUserAgentExtra(tc.env)
+			out, err := sdkv2provider.ParseUserAgentExtra(tc.env)
 
 			assert.Equal(t, tc.err, err)
 			assert.Equal(t, tc.out, out)
