+++ internal/common/provider/test_utils.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/databricks/terraform-provider-databricks/common"
 	pluginframeworkprovider "github.com/databricks/terraform-provider-databricks/internal/pluginframework/provider"
+	sdkv2provider "github.com/databricks/terraform-provider-databricks/internal/sdkv2/provider"
 	"github.com/hashicorp/terraform-plugin-framework/provider"
 	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
@@ -138,7 +139,7 @@ func (pf providerFixture) configureProviderAndReturnClient_SDKv2(t *testing.T) (
 	for k, v := range pf.env {
 		t.Setenv(k, v)
 	}
-	p := DatabricksProvider()
+	p := sdkv2provider.DatabricksProvider()
 	ctx := context.Background()
 	diags := p.Configure(ctx, terraform.NewResourceConfigRaw(pf.rawConfigSDKv2()))
 	if len(diags) > 0 {
