+++ internal/sdkv2/provider/gen/data_x.go.tmpl
