+++ internal/sdkv2/provider/gen/data_x.md.go.tmpl
