+++ internal/sdkv2/provider/gen/data_x_manual.go.tmpl
