+++ internal/sdkv2/provider/provider.go
@@ -54,7 +54,7 @@ func init() {
 	useragent.WithProduct(pluginframeworkprovider.GetProviderName(), common.Version())
 
 	userAgentExtraEnv := os.Getenv("DATABRICKS_USER_AGENT_EXTRA")
-	out, err := parseUserAgentExtra(userAgentExtraEnv)
+	out, err := ParseUserAgentExtra(userAgentExtraEnv)
 
 	if err != nil {
 		panic(fmt.Errorf("failed to parse DATABRICKS_USER_AGENT_EXTRA: %s", err))
@@ -310,7 +310,7 @@ func ConfigureDatabricksClient(ctx context.Context, d *schema.ResourceData) (any
 	return pc, nil
 }
 
-type userAgentExtra struct {
+type UserAgentExtra struct {
 	Key   string
 	Value string
 }
@@ -323,8 +323,8 @@ type userAgentExtra struct {
 // tchar = "!" / "#" / "$" / "%" / "&" / "'" / "*" / "+" / "-" / "." / "^" / "_" / "`" / "|" / "~" / DIGIT / ALPHA
 var productRegexRfc9110 = regexp.MustCompile("^([!#$%&'*+\\-.^_`|~0-9A-Za-z]+)(/([!#$%&'*+\\-.^_`|~0-9A-Za-z]+))?$")
 
-func parseUserAgentExtra(env string) ([]userAgentExtra, error) {
-	out := []userAgentExtra{}
+func ParseUserAgentExtra(env string) ([]UserAgentExtra, error) {
+	out := []UserAgentExtra{}
 
 	products := strings.FieldsFunc(env, func(r rune) bool {
 		return unicode.IsSpace(r)
@@ -341,7 +341,7 @@ func parseUserAgentExtra(env string) ([]userAgentExtra, error) {
 			return nil, fmt.Errorf("product string must include version: %s", product)
 		}
 
-		out = append(out, userAgentExtra{
+		out = append(out, UserAgentExtra{
 			Key:   match[1],
 			Value: match[3],
 		})
